Reorganize so that TreeClique.data can be BayesTreeNodeData
